requirements.txt
# Production dependencies
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0 
python-dateutil>=2.8.2
open-bus-stride-client